Fix more build errors (only 98 left outside tests!); implement SoftCache and font-related methods of GraphicsEnvironment; eliminate more dead code in sun.*

+++ src/share/classes/sun/awt/image/XbmImageDecoder.java
@@ -130,7 +130,7 @@ public class XbmImageDecoder extends ImageDecoder {
           }
           for (int mask = 1; mask <= 0x80; mask <<= 1) {
             if (x < W) {
-              raster[x] = (n & mask) != 0 ? 1 : 0;
+              raster[x] = (byte) ((n & mask) != 0 ? 1 : 0);
             }
             x++;
           }
